support streaming sparse read & upgrade mutate proto (#5)
* support streaming sparse read & upgrade proto

* remove statsd support
