+++ tablestore/ddl/index_change_test.go
@@ -106,11 +106,11 @@ func (s *testIndexChangeSuite) TestSmallFunctions(c *C) {
 	txn, err := d.GetStore().Begin()
 	c.Assert(err, IsNil)
 	ctx1 := context.WithValue(context.Background(), sessionctx.TxnIDKey, txn)
-	err = originTable.Insert(ctx1, []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow})
+	err = originTable.Insert(ctx1, &tspb.KeySet{}, "", []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow})
 	c.Assert(err, IsNil)
-	err = originTable.Insert(ctx1, []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow1})
+	err = originTable.Insert(ctx1, &tspb.KeySet{}, "", []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow1})
 	c.Assert(err, IsNil)
-	err = originTable.Insert(ctx1, []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow2})
+	err = originTable.Insert(ctx1, &tspb.KeySet{}, "", []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow2})
 	c.Assert(err, IsNil)
 	err = txn.Commit(context.Background())
 	c.Assert(err, IsNil)
@@ -154,11 +154,11 @@ func (s *testIndexChangeSuite) TestIndexChange(c *C) {
 	c.Assert(err, IsNil)
 
 	ctx1 := context.WithValue(context.Background(), sessionctx.TxnIDKey, txn)
-	err = originTable.Insert(ctx1, []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow})
+	err = originTable.Insert(ctx1, &tspb.KeySet{}, "", []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow})
 	c.Assert(err, IsNil)
-	err = originTable.Insert(ctx1, []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow1})
+	err = originTable.Insert(ctx1, &tspb.KeySet{}, "", []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow1})
 	c.Assert(err, IsNil)
-	err = originTable.Insert(ctx1, []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow2})
+	err = originTable.Insert(ctx1, &tspb.KeySet{}, "", []string{"c1", "c2"}, []*tspb.ListValue{&valueOneRow2})
 	c.Assert(err, IsNil)
 	err = txn.Commit(context.Background())
 	c.Assert(err, IsNil)
