+++ tablestore/table/tables/sparse.go
@@ -32,6 +32,7 @@ import (
 	"github.com/zhihu/zetta/pkg/model"
 	"github.com/zhihu/zetta/pkg/tablecodec"
 	"github.com/zhihu/zetta/tablestore/sessionctx"
+	"github.com/zhihu/zetta/tablestore/table"
 )
 
 var (
@@ -39,6 +40,14 @@ var (
 	fetchSparseCounterErr  = metrics.FetchSparseCounter.WithLabelValues(metrics.LblError)
 	fetchSparseDurationOK  = metrics.FetchSparseDuration.WithLabelValues(metrics.LblOK)
 	fetchSparseDurationErr = metrics.FetchSparseDuration.WithLabelValues(metrics.LblError)
+
+	batchSparseCounterOK  = metrics.BatchSparseCounter.WithLabelValues(metrics.LblOK)
+	batchSparseCounterErr = metrics.BatchSparseCounter.WithLabelValues(metrics.LblError)
+	batchSparseDurationOK = metrics.BatchSparseDuration.WithLabelValues(metrics.LblOK)
+
+	scanSparseCounterOK  = metrics.ScanSparseCounter.WithLabelValues(metrics.LblOK)
+	scanSparseCounterErr = metrics.ScanSparseCounter.WithLabelValues(metrics.LblError)
+	scanSparseDurationOK = metrics.ScanSparseDuration.WithLabelValues(metrics.LblOK)
 )
 
 type sparseRow struct {
@@ -118,11 +127,13 @@ func (t *tableCommon) prepareRowKeys(ctx context.Context, sr *tspb.SparseReadReq
 		}
 
 		srows = append(srows, srow)
-
-		if len(row.Qualifiers) == 0 {
+		qLen := len(row.Qualifiers)
+		if qLen == 0 {
 			srow.entireRow = true
+			srow.rowCells.Cells = make([]*tspb.Cell, 0, 200000)
 			continue
 		}
+		srow.rowCells.Cells = make([]*tspb.Cell, 0, qLen)
 
 		for _, col := range row.Qualifiers {
 			rkey := tablecodec.EncodePkCFColumn(rkeyPrefix, []byte(col))
@@ -138,44 +149,61 @@ func (t *tableCommon) fetchSparse(ctx context.Context, rowKeys []kv.Key, cellMap
 		wg         sync.WaitGroup
 		err1, err2 error
 	)
+	streamRead := sessionctx.StreamReadFromContext(ctx)
+	startTs := time.Now()
 
-	wg.Add(2)
-	go func() {
-		defer wg.Done()
-		err1 = t.readSparseBatch(ctx, rowKeys, cellMap, cf)
-	}()
+	if len(rowKeys) > 0 {
+		if !streamRead {
+			wg.Add(1)
+		}
+		go func() {
+			if !streamRead {
+				defer wg.Done()
+			}
+			err1 = t.readSparseBatch(ctx, rowKeys, cellMap, cf, ri)
+			if err1 != nil {
+				batchSparseCounterErr.Inc()
+				return
+			}
+			batchSparseCounterOK.Inc()
+			batchSparseDurationOK.Observe(time.Since(startTs).Seconds())
+		}()
+	}
+	err2 = t.scanSparseRow(ctx, srows, cf, ri)
+	if err2 != nil {
+		scanSparseCounterErr.Inc()
+		return err2
+	}
+	scanSparseCounterOK.Inc()
+	scanSparseDurationOK.Observe(time.Since(startTs).Seconds())
+	if !streamRead {
+		wg.Wait()
+	}
 
-	go func() {
-		defer wg.Done()
-		err2 = t.scanSparseRow(ctx, srows, cf)
-	}()
-	wg.Wait()
 	if err1 != nil {
 		return err1
 	}
-	if err2 != nil {
-		return err2
-	}
-	for _, srow := range srows {
-		ri.sendData(srow.rowCells)
+	if !streamRead {
+		for _, srow := range srows {
+			ri.sendData(srow.rowCells)
+		}
 	}
 	return nil
 }
 
-func (t *tableCommon) readSparseBatch(ctx context.Context, rowKeys []kv.Key, cellMap map[string]*sparseRow, cf *model.ColumnFamilyMeta) error {
+func (t *tableCommon) readSparseBatch(ctx context.Context, rowKeys []kv.Key, cellMap map[string]*sparseRow, cf *model.ColumnFamilyMeta, ri *resultIter) error {
 	var (
-		txn = ctx.Value(sessionctx.TxnIDKey).(kv.Transaction)
-		sc  = &stmtctx.StatementContext{TimeZone: time.Local}
-
-		preIdx = -1
+		txn                   = ctx.Value(sessionctx.TxnIDKey).(kv.Transaction)
+		sc                    = &stmtctx.StatementContext{TimeZone: time.Local}
+		streamRead            = sessionctx.StreamReadFromContext(ctx)
+		preIdx                = -1
+		preSrow    *sparseRow = nil
 	)
-
 	valsMap, err := txn.BatchGet(rowKeys)
 	if err != nil {
 		logutil.Logger(ctx).Error("batch get row value error", zap.Error(err))
 		return err
 	}
-
 	for i := 0; i < len(rowKeys); i++ {
 		val, ok := valsMap[string(rowKeys[i])]
 		if !ok {
@@ -184,7 +212,13 @@ func (t *tableCommon) readSparseBatch(ctx context.Context, rowKeys []kv.Key, cel
 
 		srow := cellMap[string(rowKeys[i])]
 		if srow.idx > preIdx {
+			if preSrow != nil && streamRead {
+				if err := ri.sendData(preSrow.rowCells); err == table.ErrUserLimitReached {
+					return nil
+				}
+			}
 			preIdx = srow.idx
+			preSrow = srow
 
 			if srow.rowCells == nil {
 				srow.rowCells = &tspb.SliceCell{
@@ -217,16 +251,17 @@ func (t *tableCommon) readSparseBatch(ctx context.Context, rowKeys []kv.Key, cel
 	return nil
 }
 
-func (t *tableCommon) scanSparseRow(ctx context.Context, srows []*sparseRow, cf *model.ColumnFamilyMeta) error {
+func (t *tableCommon) scanSparseRow(ctx context.Context, srows []*sparseRow, cf *model.ColumnFamilyMeta, ri *resultIter) error {
 	var (
-		txn = ctx.Value(sessionctx.TxnIDKey).(kv.Transaction)
-		sc  = &stmtctx.StatementContext{TimeZone: time.Local}
+		txn        = ctx.Value(sessionctx.TxnIDKey).(kv.Transaction)
+		sc         = &stmtctx.StatementContext{TimeZone: time.Local}
+		streamRead = sessionctx.StreamReadFromContext(ctx)
 	)
+
 	for _, srow := range srows {
 		if !srow.entireRow {
 			continue
 		}
-
 		iter, err := txn.Iter(srow.rkeyPrefix, srow.rkeyPrefix.PrefixNext())
 		if err != nil {
 			logutil.Logger(ctx).Error("get iterator error", zap.Error(err))
@@ -258,6 +293,11 @@ func (t *tableCommon) scanSparseRow(ctx context.Context, srows []*sparseRow, cf 
 			}
 			srow.rowCells.Cells = append(srow.rowCells.Cells, cell)
 			iter.Next()
+		}
+		if streamRead {
+			if err := ri.sendData(srow.rowCells); err != nil {
+				return err
+			}
 		}
 	}
 	return nil
